python_engine/requirements.txt

fastapi
uvicorn
google-generativeai
vtracer
pillow
python-multipart
